requirements.txt
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
plotly==5.20.0
pillow==10.3.0
soundfile==0.12.1
qutip==4.7.5
ratelimit==2.2.1
bokeh==3.4.1
matplotlib==3.8.4
scipy==1.12.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
packaging==24.0
tenacity==8.2.3
typing_extensions==4.11.0